chore(linter): extract path discovery routines (6057)
## Summary
- move path discovery out of Linter into a standalone public discovery module
- pass configured file extensions explicitly while preserving sqruff ignore traversal behavior
- relocate the existing path discovery tests alongside the extracted implementation

Ported from SQLFluff c7e791d5ff3bf681a1eb2d717a69c8e166029c42
https://github.com/sqlfluff/sqlfluff/pull/6057
https://github.com/sqlfluff/sqlfluff/commit/c7e791d5ff3bf681a1eb2d717a69c8e166029c42
