+++ .sqlfluff-sha
@@ -1,1 +1,1 @@
-cc40b83919cbfcc9eec3d223578f826f62ff72fd
+c7e791d5ff3bf681a1eb2d717a69c8e166029c42
