+++ crates/lib/src/core/linter.rs
@@ -1,4 +1,5 @@
 pub mod common;
 pub mod core;
+pub mod discovery;
 pub mod linted_file;
 pub mod linting_result;
